FOR CONSIDERATION: Boost hitbox changes
Found during playtesting that there would be frustrating moments where I'd want to adjust a boost slightly to get a different angle and distance, but the capsule hitboxes would mean the ball would be able to pass through the corner of the boost, where I needed it to hit, and not work even though the ball overlapped it. I've added box colliders and disabled the capsule colliders, so the hitbox on them more closely matches the sprite. I've put this as it's own commit so it can be easily changed back if people think the capsule was better